ADDED[Added Custom Login and and refactor views]
