+++ app/inventory/views.py
@@ -3,7 +3,7 @@ from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
 from rest_framework_simplejwt.views import TokenObtainPairView
 from rest_framework import generics, status
 from rest_framework_simplejwt.tokens import RefreshToken
-from core.custom_auth import EmailNameAuthBackend
+from .custom_auth import EmailNameAuthBackend
 from rest_framework.views import APIView
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.decorators import permission_classes
